refactor(bridge): abstract confimation monitoring logic
